Moved file related helper function to file_utils.cpp file
