+++ src/utils/file_utils.cpp
@@ -249,6 +249,26 @@ void make_directory(const std::string& path) {
 
 }
 
+void remove_multiple_slashes(std::string& str) {
+  std::stringstream s;
+  char last_char_was_slash = false;
+  for (size_t i=0; i<str.size(); i++) {
+    if (last_char_was_slash) {
+      if (str[i] != '/') {
+        s << str[i];
+      }
+    } else {
+      s << str[i];
+    }
+    if (str[i] == '/') {
+      last_char_was_slash = true;
+    } else {
+      last_char_was_slash = false;
+    }
+  }
+  str = s.str();
+}
+
 } // namespace file
 
 } // namespace lbann
